requirements.txt
streamlit
selenium
beautifulsoup4
pymongo
python-dotenv
openai
webdriver-manager